+++ sharding-jdbc-core/src/main/java/io/shardingjdbc/core/jdbc/metadata/ShardingMetaDataHandlerFactory.java
@@ -15,9 +15,13 @@
  * </p>
  */
 
-package io.shardingjdbc.core.metadata.dialect;
+package io.shardingjdbc.core.jdbc.metadata;
 
 import io.shardingjdbc.core.constant.DatabaseType;
+import io.shardingjdbc.core.jdbc.metadata.dialect.DefaultShardingMetaDataHandler;
+import io.shardingjdbc.core.jdbc.metadata.dialect.H2ShardingMetaDataHandler;
+import io.shardingjdbc.core.jdbc.metadata.dialect.MySQLShardingMetaDataHandler;
+import io.shardingjdbc.core.jdbc.metadata.dialect.ShardingMetaDataHandler;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
